Refactor TIM processing and enhance geometry handling
- Updated FieldConversions to correct zoom factor calculations for longitude and latitude conversions.
- Introduced GeometryCollection class to support mixed geometry types in TIM data frames.
- Enhanced TimConverter to handle directionality and anchor points for path and circle regions.
- Improved TimGeometryConverter to create GeometryCollection when multiple region types are present.
- Updated tests to validate new geometry handling and ensure compliance with ITWG best practices.

+++ jpo-geojsonconverter/src/main/java/us/dot/its/jpo/geojsonconverter/pojos/geojson/tim/ProcessedCircleRegionInfo.java
@@ -24,7 +24,8 @@ import lombok.extern.slf4j.Slf4j;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Slf4j
 public class ProcessedCircleRegionInfo extends ProcessedRegionInfoBase {
-    private Integer radius;
+    /** Circle radius in meters, preserving the precision provided by J2735 DistanceUnits. */
+    private Double radius;
 
     /**
      * Override to ensure only heading-based direction info is allowed for circle regions.
